Coerce return value of boost::size
The return value type of boost::size for arrays changes signedness between boost-1.49 and boost
1.5x. It is also impossible to synthesize the return type. Since we should compile against both
versions, coercing the return value to int is the simplest solution.

+++ src/mongo/db/auth/privilege_set.cpp
@@ -79,7 +79,7 @@ namespace mongo {
         resourceSearchList[1] = nsToDatabaseSubstring(desiredPrivilege.getResource());
 
         ActionSet unmetRequirements = desiredPrivilege.getActions();
-        for (int i = 0; i < boost::size(resourceSearchList); ++i) {
+        for (int i = 0; i < static_cast<int>(boost::size(resourceSearchList)); ++i) {
             ResourcePrivilegeCacheEntry* entry = _lookupEntry(resourceSearchList[i]);
             if (NULL == entry)
                 continue;
